requirements.txt
requests==2.21.0
selenium==3.141.0
beautifulsoup4==4.7.1
python-dotenv==0.10.1
